projet_ihm-main/rag-example-main/requirements.txt
langchain-community>=0.3.27,<0.4
langchain-huggingface>=1.0.0,<2
langchain-google-genai>=3.0.0,<4
langchain-text-splitters>=0.3.9,<1.0.0
sentence-transformers
chromadb>=0.5.0
streamlit
openai
pypdf
python-dotenv
fastapi
uvicorn